% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadisco.R
\name{createMetaDisco}
\alias{createMetaDisco}
\title{#' @export
createMetaDisco <- function(condition1,
condition2,
n_tile,
window_size,
k_neigh,
bandwidth,
n_bootstrap,
bootstrap_method,
weight_scheme,
n_cores = parallel::detectCores()){
list1 <- list()
for(i in 1:length(condition1)){
print(condition1\link{i})
if(is.list(condition1)){
list1[\link{i}] <- createSpomic(condition1[\link{i}])
} else{
list1[\link{i}] <- createSpomic(condition1\link{i})
}
# list1[\link{i}] <- createSpomic(condition1\link{i})
list1[\link{i}] <- setSpomicHypers(spomic_obj=list1[\link{i}],
n_tile=n_tile,
window_size=window_size,
k_neigh=k_neigh,
bandwidth=bandwidth,
n_bootstrap=n_bootstrap,
weight_scheme=weight_scheme)
list1[\link{i}] <- runCLQ(spomic_obj=list1[\link{i}],
bootstrap_method=bootstrap_method,
n_cores=n_cores)
}}
\usage{
createMetaDisco(
  condition1,
  condition2,
  tile_size,
  window_size,
  k_neigh,
  bandwidth,
  n_bootstrap,
  weight_scheme,
  n_cores = parallel::detectCores()
)
}
\arguments{
\item{condition1}{A list of \code{Spomic} objects, data frames of paths to CSV files for the second condition.}

\item{tile_size}{A numeric value specifying the size of tiles to be used for spatial bootstrapping.}

\item{window_size}{An integer specifying the size of the spatial window for analysis.}

\item{k_neigh}{An integer representing the number of nearest neighbors to consider in colocalization analysis.}

\item{bandwidth}{An integer defining the bandwidth used to restrict colocalization analysis.}

\item{n_bootstrap}{An integer specifying the number of bootstrapping iterations to perform.}

\item{weight_scheme}{A character string describing the weighting scheme for colocalization.}

\item{n_cores}{An integer specifying the number of CPU cores to use for parallel processing.
Defaults to the number of available cores.}
}
\value{
A \code{MetaDisco} object containing processed spatial omics data for both conditions, along with the analysis parameters.
}
\description{
Constructs a \code{MetaDisco} object from two conditions, each containing one or
more \code{Spomic} objects or data respresentations that can be converted into \code{Spomic}
objects. This function processes spatial omics data for both conditions, sets
analysis parameters (hypers), and performs colocalization analysis with
optional bootstrapping.
}
\details{

}
