% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spomic.R
\name{createSpomic}
\alias{createSpomic}
\title{Create a Spomic object}
\usage{
createSpomic(p, drop_na = TRUE)
}
\arguments{
\item{p}{A data frame or a path to a CSV file.}

\item{drop_na}{A boolean for whether to remove rows containing NAs. Defaults to TRUE.}
}
\value{
An object of class \code{Spomic}.
}
\description{
Constructs a 'Spomic' object from either a data frame or a path to a CSV file.
}
\details{
The input data must include the following columns:
\itemize{
\item \code{x}: Numeric values representing sptial x-coordinates
\item \code{y}: Numeric values representing spatial y-coordinates.
\item \code{cell_type}: A character vector denoting cell type assignments.
\item \code{sample}: A character vector specifying the sample name / ID.
}
}
\examples{
# Example with a data frame
df <- data.frame(
  x = runif(10),
  y = runif(10),
  cell_type = sample(c("A", "B", "C"), 10, replace = TRUE),
  sample = "Sample_1"
)
spomic_obj <- createSpomic(df)

# Example with a CSV file
# write.csv(df, "data.csv", row.names = FALSE)
# spomic_obj <- createSpomic("data.csv")

}
