% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{filterTiles}
\alias{filterTiles}
\title{Filter tiled data}
\usage{
filterTiles(tibbles, find_empty = FALSE)
}
\arguments{
\item{tibbles}{Filters a list of tibbles based on the presence or absence of \code{NA} values.}

\item{find_empty}{Logical. If \code{TRUE}, returns only tiles that contain \code{NA} values.
If \code{FALSE}, returns only tiles without any \code{NA} values. Default is \code{FALSE}.}
}
\value{
A filtered list of tibbles based on the specified condition.
}
\description{
Filters a list of tibbles based on the presence or absence of \code{NA} values.
}
