% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization_functions.R
\name{getNeighbors}
\alias{getNeighbors}
\title{Get nearest neighbors for a Spomic object}
\usage{
getNeighbors(spomic_obj)
}
\arguments{
\item{spomic_obj}{A \code{Spomic} object containing spatial coordinates.}
}
\value{
A list where each element represents the neighbors for a given point,
indexed by row number.
}
\description{
Computes the nearest neighbors for each spatial point in a \code{Spomic} object.
}
\details{
The function retrieves the nearest \code{k} neighbors for each spatial point in \code{spomic@coords}.
}
