% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spomic.R
\name{saveSpomic}
\alias{saveSpomic}
\title{Save a Spomic object}
\usage{
saveSpomic(spomic_obj, path)
}
\arguments{
\item{spomic_obj}{An object of class \code{Spomic}.}

\item{path}{A character string specifying the directory where the object should be saved.}
}
\description{
Saves a \code{Spomic} object as an RDS file in a specified directory. The RDS will save in a subdirectory containing hyperparameter information.
}
\details{
The function constructs a directory path using metadata from the \code{Spomic} object,
including parameters such as \code{tilesize}, \code{nbors} (neighbors), \code{bandwidth},
\code{bootstrap}, and \code{window}. The directory and file are created dynamically if
they do not exist.

The final file is saved with the naming convention \verb{<sample_id>_spomic.Rds}.
}
