% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization_functions.R
\name{searchCellNeighbors}
\alias{searchCellNeighbors}
\title{Search for cell neighbors}
\usage{
searchCellNeighbors(spomic_obj, cellA, cellB)
}
\arguments{
\item{spomic_obj}{A \code{Spomic} object.}

\item{cellA}{A character string specifying the type of cell (e.g., "A") to search for as the primary cell.}

\item{cellB}{A character string specifying the type of cell (e.g., "B") to evaluate as neighbors of \code{cellA}.}
}
\value{
A matrix where rows correspond to cells of type \code{cellA} and the single column contains the weighted
proportion of neighbors of type \code{cellB}.
If no cells of type \code{cellA} are found, the function returns \code{NULL}.
}
\description{
Computes the proportion of neighbors of type \code{cellB} for each cell of type \code{cellA }
in a \code{Spomic} object. This function supports different weighting schemes for calculating neighbor proportions.
}
\details{
The function uses the \verb{@nb_list} slot in the \code{Spomic} object to retrieve the
neighbors for each cell of type \code{cellA}.
Proportions are weighted based on the distance to neighbors, using the following schemes:
\itemize{
\item \code{"constant"}: All neighbors are weighted equally.
\item \code{"linear"}: Neighbors closer to the current cell are weighted higher, and the weights decay linearly.
}
}
