% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spomic.R
\name{setSpomicHypers}
\alias{setSpomicHypers}
\title{Set hyperparameters for a Spomic object}
\usage{
setSpomicHypers(
  spomic_obj,
  tile_size,
  window_size,
  k_neigh,
  bandwidth,
  n_bootstrap,
  weight_scheme,
  precompute_neighbors = TRUE
)
}
\arguments{
\item{spomic_obj}{An object of class Spomic.}

\item{tile_size}{Tile size.}

\item{window_size}{Window size.}

\item{k_neigh}{Number of neighbors.}

\item{bandwidth}{Bandwidth parameter.}

\item{n_bootstrap}{Number of bootstrap samples.}

\item{weight_scheme}{Weight scheme for colocation quotient calculation ("constant", "linear", "squared_exponential")}

\item{precompute_neighbors}{Boolean indicating whether to precompute neighbors.}
}
\value{
Updated Spomic object.
}
\description{
Set hyperparameters for a Spomic object
}
