% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spomic.R
\docType{class}
\name{Spomic-class}
\alias{Spomic-class}
\title{Spomic (Spatial omic) Class}
\description{
Defines the structure for Spomic objects, used to store spatial and colocalization data.
}
\section{Slots}{

\describe{
\item{\code{sample_id}}{A string with the sample name / ID}

\item{\code{df}}{A data frame containing columns for coordinates, sample, and cell type.}

\item{\code{coords}}{A data frame specifying spatial coordinates (TODO: in the future, this can probably be removed...)}

\item{\code{cell_assignments}}{A vector assigning cells to groups or types.}

\item{\code{tile_size}}{Numeric, tile size to be used in spatial bootstrapping.}

\item{\code{window_size}}{Integer, window size for spatial analysis.}

\item{\code{k_neigh}}{Integer, number of nearest neighbors to consider for colocalization calculation}

\item{\code{bandwidth}}{Integer, bandwidth for smoothing operations.}

\item{\code{n_bootstrap}}{Integer, number of spatial bootstrapping iterations to perform.}

\item{\code{nb_list}}{A list of indices for the k_neigh nearest neighbors of each cell.}

\item{\code{colocalization}}{A data frame with colocalization data processed for Gaussianity and bias.}

\item{\code{weight_scheme}}{A character string describing the weighting scheme used in colocalization calculation.}
}}
