% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{tileImage}
\alias{tileImage}
\title{Tile a 2D image}
\usage{
tileImage(img, tile_size)
}
\arguments{
\item{img}{A data frame with numeric \code{x} and \code{y} columns representing coordinates.}

\item{tile_size}{A positive numeric value specifying the size of each tile.}
}
\value{
A list of tibbles, where each tibble represents a tile with rescaled coordinates.
}
\description{
Splits a 2D dataset with \code{x} and \code{y} coordinates into tiles of specified size
and rescales the coordinates within each tile to the range \link{0, 1}.
}
\details{
The function calculates tile row and column indices based on the \code{tile_size},
assigns each data point to a corresponding tile, and rescales te coordinates
within each tile  to a normalized scale between 0 and 1.
}
