% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{windowBootstrap}
\alias{windowBootstrap}
\title{Perform Window Bootstrap}
\usage{
windowBootstrap(spomic_obj)
}
\arguments{
\item{spomic_obj}{A \code{Spomic} object containing the spatial data and tiling parameters set using \code{metadisco::setSpomicHypers()}}
}
\value{
A new bootstrapped \code{Spomic} object where the data has been resampled.
}
\description{
Perform Window Bootstrap
}
\details{
The function performs the following steps:
\enumerate{
\item Tiles the data into smaller chunks based on the tilesize parameter.
\item Identifies empty and non-empty tiles.
\item For each non-empty tile, selects neighboring tiles within the specified \code{window_size}.
\item Samples tile from the neighboring non-empty tiles.
\item Generate new image and returns a new bootstrapped \code{Spomic} object.
}
}
