% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadisco.R
\docType{class}
\name{MetaDisco-class}
\alias{MetaDisco-class}
\title{MetaDisco Class}
\description{
Defines the structure for the \code{MetaDisco} class, which is used to represent metadata
and analysis parameters for comparative spatial omics studies. This class
supports the storage of confitions, spatial tiling parameters, windowing,
bootstrap configurations, and analysis results
}
\section{Slots}{

\describe{
\item{\code{condition1}}{A list of \code{Spomic} classes for samples within the first group/condition.}

\item{\code{condition2}}{A list of \code{Spomic} classes for samples within the second group/condition.}

\item{\code{tile_size}}{A numeric value indicating the size of tiles to be used in the analysis and bootstrapping.}

\item{\code{window_size}}{An integer specifying the size of the spatial window for analysis.}

\item{\code{k_neigh}}{An integer representing the number of nearest neighbors to consider
in the colocalization calculations.}

\item{\code{bandwidth}}{An integer specifying the maximum distance from a point to consider in colocalization calculations.}

\item{\code{n_bootstrap}}{An integer denoting the number of bootstrapping iterations to perform.}

\item{\code{weight_scheme}}{A character string describing the weighting scheme to be applied for colocalization calculations.}

\item{\code{results}}{A data frame containing the results of differential colocalization testing.}
}}
